+++ src/main/java/kr/modusplant/modules/signup/normal/app/controller/NormalSignUpController.java
@@ -1,28 +1,20 @@
-package kr.modusplant.modules.signup.normal.controller;
+package kr.modusplant.modules.signup.normal.app.controller;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.exc.UnrecognizedPropertyException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import kr.modusplant.domains.member.domain.model.SiteMember;
-import kr.modusplant.domains.member.domain.model.SiteMemberAuth;
-import kr.modusplant.domains.member.domain.model.SiteMemberTerm;
-import kr.modusplant.domains.member.domain.service.supers.SiteMemberAuthCrudService;
-import kr.modusplant.domains.member.domain.service.supers.SiteMemberCrudService;
-import kr.modusplant.domains.member.domain.service.supers.SiteMemberTermCrudService;
-import kr.modusplant.domains.member.enums.AuthProvider;
 import kr.modusplant.domains.term.domain.model.Term;
 import kr.modusplant.global.app.servlet.response.DataResponse;
-import kr.modusplant.modules.signup.normal.model.request.NormalSignUpRequest;
-import kr.modusplant.modules.signup.normal.service.NormalSignUpApplicationService;
+import kr.modusplant.modules.signup.normal.app.http.request.NormalSignUpRequest;
+import kr.modusplant.modules.signup.normal.app.service.NormalSignUpApplicationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.dao.InvalidDataAccessResourceUsageException;
 import org.springframework.http.ResponseEntity;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -37,10 +29,6 @@ import java.util.Map;
 public class NormalSignUpController {
 
     private final NormalSignUpApplicationService normalSignUpApplicationService;
-
-    private final SiteMemberTermCrudService siteMemberTermCrudService;
-    private final SiteMemberAuthCrudService siteMemberAuthCrudService;
-    private final SiteMemberCrudService siteMemberCrudService;
 //    private final PasswordEncoder passwordEncoder;
 
     @Operation(
@@ -89,22 +77,18 @@ public class NormalSignUpController {
     })
     @PostMapping("/api/members/register")
     public ResponseEntity<DataResponse<Void>> saveMember(@RequestBody NormalSignUpRequest memberData) {
-        log.info("SignUpData: {}", memberData);
 
         try {
-
             if(memberData.pw().equals(memberData.pw_check())) {
-                insertMember(memberData);
+                normalSignUpApplicationService.insertMember(memberData);
                 DataResponse<Void> successDataResponse = DataResponse.of(200, "sign up successfully");
 
                 return ResponseEntity.ok(successDataResponse);
-
             } else {
                 DataResponse<Void> errorDataResponse = DataResponse.of(400, "pw and pw_check not equivalent");
 
                 return ResponseEntity.ok(errorDataResponse);
             }
-
         } catch (Exception e) {
             String exceptionMessage = "";
 
@@ -126,30 +110,4 @@ public class NormalSignUpController {
 
     }
 
-    @Transactional
-    private void insertMember(NormalSignUpRequest memberData) {
-        SiteMember siteMember = SiteMember.builder()
-                .nickname(memberData.nickname())
-                .build();
-        SiteMember savedMember = siteMemberCrudService.insert(siteMember);
-
-        SiteMemberAuth siteMemberAuth = SiteMemberAuth.builder()
-                .uuid(savedMember.getUuid())
-                .activeMemberUuid(savedMember.getUuid())
-                .originalMemberUuid(savedMember.getUuid())
-                .email(memberData.email())
-//                .pw(passwordEncoder.encode(requestNode.get("pw")))
-                .pw(memberData.pw())
-                .provider(AuthProvider.BASIC)
-                .build();
-        siteMemberAuthCrudService.insert(siteMemberAuth);
-
-        SiteMemberTerm siteMemberTerm = SiteMemberTerm.builder()
-                .uuid(savedMember.getUuid())
-                .agreedTermsOfUseVersion(memberData.agreedTermsOfUseVerion())
-                .agreedPrivacyPolicyVersion(memberData.agreedPrivacyPolicyVerion())
-                .agreedAdInfoReceivingVersion(memberData.agreedAdInfoRecevingVerion())
-                .build();
-        siteMemberTermCrudService.insert(siteMemberTerm);
-    }
 }
