+++ src/main/java/kr/modusplant/modules/signup/normal/mapper/domain/SiteMemberTermDomainMapper.java
@@ -1,0 +1,17 @@
+package kr.modusplant.modules.signup.normal.mapper.domain;
+
+import kr.modusplant.domains.member.domain.model.SiteMember;
+import kr.modusplant.domains.member.domain.model.SiteMemberTerm;
+import kr.modusplant.modules.signup.normal.app.http.request.NormalSignUpRequest;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.NullValuePropertyMappingStrategy;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.SET_TO_NULL,
+        unmappedTargetPolicy = ReportingPolicy.IGNORE)
+public interface SiteMemberTermDomainMapper {
+
+    @Mapping(target = "uuid", source = "siteMember.uuid")
+    SiteMemberTerm toSiteMemberTerm(NormalSignUpRequest request, SiteMember siteMember);
+}
